Fix target assembly path resolution
Don't know what changed to make the GetExecutingAssembly() return
an assembly with an empty Location property, but it did and thus
made the AssemblyRunner.WithoutAppDomain fail its file-exists check
even though we had a valid assembly.

There's no way to pass a fully qualified Assembly instance to the
XUnit AssemblyRunner, only a path string.

Reworked the target assembly resolution and so it now passes a full path
to the default assembly correctly and there are no errors.

However it still doesn't detect any tests.