Brute force assembly resolution
The assemblies would continue to not be located correctly even
though a fully qualified path was passed to the AssemblyRunner.

I attached an event handler to the app domain's AssemblyResolve event
to retry locating the assembly. Cudos to this article:
https://weblog.west-wind.com/posts/2016/dec/12/loading-net-assemblies-out-of-seperate-folders

Trying to run the tests/ project assemblies still complained about
xunit.core and xunit.assert to be missing. Adding the CopyLocalLockFileAssemblies flag
to those projects fixed the issue.