Godot 4 port (#20)
* Gitignore the .godot dir

* Godot 4 auto-upgrade

Opened the project in godot 4.1 and allowed the auto upgrade to do
its thing.

* Opening and saving the project

Godot updated the SDK attribute in the GodotXUnit.csproj file.
Some resources updated as well.

* Update target framework to net6.0 and godot SDK

* Fix changes with Vector2

* Fix Transform2D changes

* Fix mouse button enum changes

* FindNode is not called FindChild

* Fix enum to Variant casts

* Fix Variant to String conversions

* Fix PackedScene Instantiate call

* Fix the UI querying the viewport size

* Fix remaining Variant errors

* CanvasItem.Update is now called QueueRedraw

* Fix auto-convert mistake with Path

It replaced occurrences of Path with Path3D
(thinking it was the Godot node type Path).

In fact it's System.IO.Path.

* Fix override _Process method signatures

* Button.Pressed property is called ButtonPressed

* Godot.OS.Execute changed signature

* Fix target assembly OptionButton items

* Fix TreeItem child iteration

* Fix CharacterBody2D changes

CharacterBody2D now has a build in Velocity property which we'll use.
UpDirection is also defaulted to Vector2.Up

* Fix the file system operations

* Missing partial keyword

* Add MustBeVariant attribute to generic type arg

* Signal delegate names must end with EventHandler

* Update langversion to 10

* Godot changed a project.godot entry after passing build

* Removed conflicting reference causing build warning

* Fix target assembly options not being cleared

* Copy assemblies to fix xunit.core not being found

Trying to run the plugin gave the error below.

The tag CopyLocalLockFileAssemblies seems to fix the error, as the
xunit.core.dll is included with the output files.
```
Godot Engine v4.1.beta2.mono.official (c) 2007-present Juan Linietsky, Ariel Manzur & Godot Contributors.
  modules/gltf/register_types.cpp:72 - Blend file import is enabled in the project settings, but no Blender path is configured in the editor settings. Blend files will not be imported.
--- Debug adapter server started ---
--- GDScript language server started ---
  modules/mono/glue/runtime_interop.cpp:1325 - System.IO.FileNotFoundException: Could not load file or assembly 'xunit.core, Version=2.4.1.0, Culture=neutral, PublicKeyToken=8d05b1bb7a6fdb6c'. The system cannot find the file specified.
  File name: 'xunit.core, Version=2.4.1.0, Culture=neutral, PublicKeyToken=8d05b1bb7a6fdb6c'
     at System.ModuleHandle.ResolveType(QCallModule module, Int32 typeToken, IntPtr* typeInstArgs, Int32 typeInstCount, IntPtr* methodInstArgs, Int32 methodInstCount, ObjectHandleOnStack type)
     at System.ModuleHandle.ResolveTypeHandle(Int32 typeToken, RuntimeTypeHandle[] typeInstantiationContext, RuntimeTypeHandle[] methodInstantiationContext)
     at System.Reflection.RuntimeModule.ResolveType(Int32 metadataToken, Type[] genericTypeArguments, Type[] genericMethodArguments)
     at System.Reflection.CustomAttribute.FilterCustomAttributeRecord(MetadataToken caCtorToken, MetadataImport& scope, RuntimeModule decoratedModule, MetadataToken decoratedToken, RuntimeType attributeFilterType, Boolean mustBeInheritable, ListBuilder`1& derivedAttributes, RuntimeType& attributeType, IRuntimeMethodInfo& ctorWithParameters, Boolean& isVarArg)
     at System.Reflection.CustomAttribute.AddCustomAttributes(ListBuilder`1& attributes, RuntimeModule decoratedModule, Int32 decoratedMetadataToken, RuntimeType attributeFilterType, Boolean mustBeInheritable, ListBuilder`1 derivedAttributes)
     at System.Reflection.CustomAttribute.GetCustomAttributes(RuntimeMethodInfo method, RuntimeType caType, Boolean inherit)
     at System.Reflection.RuntimeMethodInfo.GetCustomAttributes(Boolean inherit)
     at Godot.Bridge.ScriptManagerBridge.UpdateScriptClassInfo(IntPtr scriptPtr, godot_string* outClassName, godot_bool* outTool, godot_bool* outGlobal, godot_string* outIconPath, godot_array* outMethodsDest, godot_dictionary* outRpcFunctionsDest, godot_dictionary* outEventSignalsDest, godot_ref* outBaseScript) in /root/godot/modules/mono/glue/GodotSharp/GodotSharp/Core/Bridge/ScriptManagerBridge.cs:line 698
```

* Fix target assembly path resolution

Don't know what changed to make the GetExecutingAssembly() return
an assembly with an empty Location property, but it did and thus
made the AssemblyRunner.WithoutAppDomain fail its file-exists check
even though we had a valid assembly.

There's no way to pass a fully qualified Assembly instance to the
XUnit AssemblyRunner, only a path string.

Reworked the target assembly resolution and so it now passes a full path
to the default assembly correctly and there are no errors.

However it still doesn't detect any tests.

* Brute force assembly resolution

The assemblies would continue to not be located correctly even
though a fully qualified path was passed to the AssemblyRunner.

I attached an event handler to the app domain's AssemblyResolve event
to retry locating the assembly. Cudos to this article:
https://weblog.west-wind.com/posts/2016/dec/12/loading-net-assemblies-out-of-seperate-folders

Trying to run the tests/ project assemblies still complained about
xunit.core and xunit.assert to be missing. Adding the CopyLocalLockFileAssemblies flag
to those projects fixed the issue.

* Auto-format on file with few changes

* Idle frame is called process_frame

* Make tests run via the main Godot thread

Godot only allows specific threads to access Node as a safety
precaution. This makes most tests always fail since XUnit spawns at
least one thread to run tests in (even if parallel is disabled)

To get around this, all tests are now invoked inside a deferred lambda
that will run on Godot's main thread. A semaphore is then used
to let XUnit wait for the test to finish.